Post Login Functionality
I used Cat's fork to implement functionality on the post-login screen. For now, there are placeholders and the js file is used for a majority of the commands. I have code hidden to use the PHP files, but I couldn't use them to test on Live Preview or my browser.